Move location of build_scenario.
